requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
tokenizers>=0.19.0
datasets>=2.18.0
accelerate>=0.28.0
onnx>=1.15.0
onnxruntime>=1.17.0
onnxruntime-gpu>=1.17.0
optimum>=1.18.0
matplotlib>=3.8.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.9.0
albumentations>=1.3.0
tqdm>=4.66.0
wandb>=0.16.0
nltk>=3.8.0
vllm>=0.4.0
qwen-vl-utils>=0.0.8
